requirements.txt
aiofiles==0.6.0
asyncpg==0.22.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
databases==0.4.1
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
httptools==0.1.1
idna==3.1
Jinja2==2.11.3
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
python-slugify==4.0.1
PyYAML==5.4.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.0b3
starlette==0.13.6
text-unidecode==1.3
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
